Added client and changed database. Also fixed bug

+++ controller/teams_manager.py
@@ -69,7 +69,7 @@ class TeamsManager:
 
     def delete_team(self, team_name, admin_email):
         if team_name in self.teams.keys():
-            if self.teams[team_name].admin_email == admin_email:
+            if self.teams[team_name].admin_email == admin_email and len(self.teams[team_name].members) == 1:
                 del self.teams[team_name]
                 return self.teams_db.remove_team(team_name)
 
@@ -77,7 +77,7 @@ class TeamsManager:
 
         else:
             team = self.teams_db.get_team(team_name)
-            if team is not None and team.admin_email == admin_email:
+            if team is not None and team.admin_email == admin_email and len(team.members) == 1:
                 return self.teams_db.remove_team(team_name)
 
             return False
